custom response and request , add api branh , movie , user , add filter

+++ goldenticketnew/src/main/java/com/example/goldenticketnew/service/ticket/ITicketService.java
@@ -5,5 +5,5 @@ import com.example.goldenticketnew.dtos.TicketDto;
 import java.util.List;
 
 public interface ITicketService {
-    List<TicketDto> getTicketsByUserId(Integer userId);
+    List<TicketDto> getTicketsByUserId(Long userId);
 }
